change getting of SyncQueue, DataInserter instances
